pyproject.toml
[tool.poetry]
name = "n5-assessment"
version = "0.1.0"
description = ""
authors = ["Ider Delzo <soloidx@gmail.com>"]
readme = "README.md"
packages = [{include = "n5_assessment"}]

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.1.6"
django-environ = "^0.9.0"
django-oauth-toolkit = "^2.2.0"
django-ninja = "^0.20.0"
pydantic = {extras = ["email"], version = "^1.10.4"}
psycopg2-binary = "^2.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
